updated update and delete function on various forms I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE I WENT INSANE

+++ admin/PasKlasAdmin.pas
@@ -19,11 +19,17 @@ type
     DBGrid1: TDBGrid;
     edtId: TEdit;
     edtHapus: TButton;
+    btnClr: TButton;
+    btnRefresh: TButton;
+    Label4: TLabel;
+    edtCari: TEdit;
     procedure btnSimpanClick(Sender: TObject);
     procedure DBGrid1CellClick(Column: TColumn);
     procedure btnUbahClick(Sender: TObject);
     procedure edtHapusClick(Sender: TObject);
     procedure FormCreate(Sender: TObject);
+    procedure btnClrClick(Sender: TObject);
+    procedure edtCariChange(Sender: TObject);
   private
     { Private declarations }
   public
@@ -42,16 +48,30 @@ uses
 
 procedure TfrKlasAdmin.btnSimpanClick(Sender: TObject);
 begin
-modulDB.ZqKlasAdmin.SQL.Clear;
-modulDB.ZqKlasAdmin.SQL.Text := '';
-modulDB.ZqKlasAdmin.SQL.Text := 'INSERT INTO klasifikasi (nama, nomor, status) VALUES (:nama, :nomor, :status)';
+if (edtNama.Text := '') or (edtNo.Text := '') or (cbxSts.Text := '') or (cbxSts.Text := '') then
+  begin
+    ShowMessage('Ada Data Yang Kosong!');
+  end
+  else
+  begin
+    if MessageDlg('Apa Anda yakin ingin MENAMBAHKAN data?', mtConfirmation, [mbYes, mbNo], 0) = mrYes then
+      begin
+        modulDB.ZqKlasAdmin.SQL.Clear;
+        modulDB.ZqKlasAdmin.SQL.Text := '';
+        modulDB.ZqKlasAdmin.SQL.Text := 'INSERT INTO klasifikasi (nama, nomor, status) VALUES (:nama, :nomor, :status)';
 
-modulDB.ZqKlasAdmin.ParamByName('nama').Value := edtNama.Text;
-modulDB.ZqKlasAdmin.ParamByName('nomor').Value := edtNo.Text;
-modulDB.ZqKlasAdmin.ParamByName('status').Value := cbxSts.Text;
+        modulDB.ZqKlasAdmin.ParamByName('nama').Value := edtNama.Text;
+        modulDB.ZqKlasAdmin.ParamByName('nomor').Value := edtNo.Text;
+        modulDB.ZqKlasAdmin.ParamByName('status').Value := cbxSts.Text;
 
-modulDB.ZqKlasAdmin.ExecSQL;
-modulDB.DsKlasAdmin.DataSet.Refresh;
+        modulDB.ZqKlasAdmin.ExecSQL;
+        modulDB.DsKlasAdmin.DataSet.Refresh;
+      end
+      else
+      begin
+        ShowMessage('Penambahan Data Di Batalkan.');
+      end;
+  end;
 end;
 
 procedure TfrKlasAdmin.DBGrid1CellClick(Column: TColumn);
@@ -64,27 +84,55 @@ end;
 
 procedure TfrKlasAdmin.btnUbahClick(Sender: TObject);
 begin
-modulDB.ZqKlasAdmin.SQL.Clear;
-modulDB.ZqKlasAdmin.SQL.Text := '';
-modulDB.ZqKlasAdmin.SQL.Text := 'UPDATE klasifikasi SET nama = :nama, nomor = :nomor, status = :status WHERE id = :id';
+if (edtId.Text := '') or (edtNama.Text := '') or (edtNo.Text := '') then
+  begin
+    ShowMessage('Data Yang Akan Di UBAH Belum Di Pilih!');
+  end
+  else
+  begin
+    if MessageDlg('Apa Anda yakin ingin MENGUBAH data?', mtConfirmation, [mbYes, mbNo], 0) = mrYes then
+      begin
+        modulDB.ZqKlasAdmin.SQL.Clear;
+        modulDB.ZqKlasAdmin.SQL.Text := '';
+        modulDB.ZqKlasAdmin.SQL.Text := 'UPDATE klasifikasi SET nama = :nama, nomor = :nomor, status = :status WHERE id = :id';
 
-modulDB.ZqKlasAdmin.ParamByName('id').Value := edtId.Text;
-modulDB.ZqKlasAdmin.ParamByName('nama').Value := edtNama.Text;
-modulDB.ZqKlasAdmin.ParamByName('nomor').Value := edtNo.Text;
-modulDB.ZqKlasAdmin.ParamByName('status').Value := cbxSts.Text;
+        modulDB.ZqKlasAdmin.ParamByName('id').Value := edtId.Text;
+        modulDB.ZqKlasAdmin.ParamByName('nama').Value := edtNama.Text;
+        modulDB.ZqKlasAdmin.ParamByName('nomor').Value := edtNo.Text;
+        modulDB.ZqKlasAdmin.ParamByName('status').Value := cbxSts.Text;
 
-modulDB.ZqKlasAdmin.ExecSQL;
-modulDB.DsKlasAdmin.DataSet.Refresh;
+        modulDB.ZqKlasAdmin.ExecSQL;
+        modulDB.DsKlasAdmin.DataSet.Refresh;
+      end
+      else
+      begin
+        ShowMessage('Pengubahan Data Di Batalkan.');
+      end;
+  end;
 end;
 
 procedure TfrKlasAdmin.edtHapusClick(Sender: TObject);
 begin
-modulDB.ZqKlasAdmin.SQL.Clear;
-modulDB.ZqKlasAdmin.SQL.Text := '';
-modulDB.ZqKlasAdmin.SQL.Text := 'DELETE FROM klasifikasi where id = :id';
-modulDB.ZqKlasAdmin.ParamByName('id').Value := edtId.Text;
-modulDB.ZqKlasAdmin.ExecSQL;
-modulDB.DsKlasAdmin.DataSet.Refresh;
+if (edtId.Text := '') then
+  begin
+    ShowMessage('Data Yang Akan Di HAPUS Belum Di Pilih!');
+  end
+  else
+  begin
+    if MessageDlg('Apa Anda yakin ingin MENGHAPUS data?', mtConfirmation, [mbYes, mbNo], 0) = mrYes then
+      begin
+        modulDB.ZqKlasAdmin.SQL.Clear;
+        modulDB.ZqKlasAdmin.SQL.Text := '';
+        modulDB.ZqKlasAdmin.SQL.Text := 'DELETE FROM klasifikasi where id = :id';
+        modulDB.ZqKlasAdmin.ParamByName('id').Value := edtId.Text;
+        modulDB.ZqKlasAdmin.ExecSQL;
+        modulDB.DsKlasAdmin.DataSet.Refresh;
+      end
+      else
+      begin
+        ShowMessage('Penghapusan Di Batalkan.');
+      end;
+  end;
 end;
 
 procedure TfrKlasAdmin.FormCreate(Sender: TObject);
@@ -94,6 +142,44 @@ DBGrid1.Columns[0].Width := 32;
 DBGrid1.Columns[1].Width := 150;
 DBGrid1.Columns[2].Width := 64;
 DBGrid1.Columns[3].Width := 64;
+end;
+
+procedure TfrKlasAdmin.btnClrClick(Sender: TObject);
+begin
+edtId.Text := '';
+edtNama.Text := '';
+edtNo.Text := '';
+cbxSts.Text := '----';
+end;
+
+procedure refreshData();
+begin
+modulDB.ZqKlasAdminView.SQL.Clear;
+modulDB.ZqKlasAdminView.SQL.Text := '';
+modulDB.ZqKlasAdminView.SQL.Text := 'SELECT * FROM klasifikasi';
+modulDB.ZqKlasAdminView.Open;
+modulDB.DsKlasAdmin.DataSet.Refresh;
+end;
+
+procedure TfrKlasAdmin.edtCariChange(Sender: TObject);
+var
+  cari : string;
+begin
+if (edtCari.Text = '') then
+  begin
+    //jika kolom pencarian kosong, kembalikan ZqDivAdminView seperti semula
+    refreshData;
+  end
+  else
+  begin
+    //jika kolom pencarian ada isinya, lakukan pencarian secara keseluruhan
+    cari := edtCari.Text;
+    modulDB.ZqKlasAdminView.SQL.Clear;
+    modulDB.ZqKlasAdminView.SQL.Text := '';
+    modulDB.ZqKlasAdminView.SQL.Text := 'SELECT * FROM klasifikasi WHERE nama LIKE ''%' +cari+ '%'' OR nomor LIKE ''%' +cari+ '%'' OR status LIKE ''%' +cari+ '%''';
+    modulDB.ZqKlasAdminView.Open;
+    modulDB.DsKlasAdmin.DataSet.Refresh;
+  end;
 end;
 
 end.
